+++ container-engine-lib/docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Features
+* Verify that, when starting a container with `shouldPublishAllPorts` == `true`, each used port gets exactly one host machine port binding
+
 ### Fixes
 * Fixed ports not getting bound when running a container with an `EXPOSE` directive in the image Dockerfile
 * Fixed broken CI